.2333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333333
first commit by vaishnavi

+++ employee_microservice/src/main/java/com/cjc/loanapplication/model/Employee.java
@@ -8,6 +8,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -20,11 +21,17 @@ public class Employee {
 @Id
 @GeneratedValue(strategy = GenerationType.AUTO)
 private Integer employeeId;
+//@NotEmpty(message = "")
 private String employeeName;
+//@NotEmpty(message = "")
 private String userName;
+//@NotEmpty(message = "")
 private String password;
+//@NotEmpty(message = "")
 private String employeeMail;
-private Integer employeeContact;
+//@NotEmpty(message = "")
+private Long employeeContact;
+//@NotEmpty(message = "")
 @Enumerated(EnumType.STRING)
 private UserType userType;
